Allow fixing version using environment variable
The latest version turns out to crash immediately when the first player connects. It would be handy in such situations to be able to fix the server to a specific version until the problem is resolved.

This change adds an environment variable `BEDROCK_DOWNLOAD_URL`. If it is set, it is used as the version to run. I created this change through the GitHub interface. Really, the readme should also be updated.